Add valid protocols to vulnerabilities, fix socket creation
All vulnerabilites now have the valid protocols that the vulnerability
can appear on.

Add timeout handling in socket creation in vulnerability scanning and
SSLvX version scanning.
