+++ SSLTest/src/scan_vulnerabilities/VulnerabilityTest.py
@@ -27,7 +27,9 @@ class VulnerabilityTest(ABC):
         :rtype: bool
         """
         logging.info(f'Testing for {self.test_name}...')
-        to_test_protocols = filter(lambda protocol: protocol in self.valid_protocols, self.supported_protocols)
+        to_test_protocols = list(filter(lambda protocol: protocol in self.valid_protocols, self.supported_protocols))
+        if len(to_test_protocols) == 0:
+            return False
         binary_protocol_versions = list(map(lambda p: self.protocol_binary_version[p], to_test_protocols))
         output = self.multithreading_tests(self.test, binary_protocol_versions)
         logging.info(f'{self.test_name} test done.')
