+++ SSLTest/src/scan_vulnerabilities/tests/InsecureRenegotiation.py
@@ -8,7 +8,7 @@ class InsecureRenegotiation(VulnerabilityTest):
     def __init__(self, supported_protocols, address):
         super().__init__(supported_protocols, address)
         self.test_name = 'Insecure Renegotiation'
-        self.valid_protocols = ['TLSv1.2']
+        self.valid_protocols = ['TLSv1.2', 'TLSv1.1', 'TLSv1.0', 'SSLv3']
         self.renegotiation_extension = bytes([
             0xff, 0x01, 0x00, 0x01, 0x00
         ])
