+++ SSLTest/src/scan_vulnerabilities/tests/SessionTicketSupport.py
@@ -7,7 +7,7 @@ class SessionTicketSupport(VulnerabilityTest):
     def __init__(self, supported_protocols, address):
         super().__init__(supported_protocols, address)
         self.test_name = 'Session Ticket Support'
-        self.valid_protocols = ['TLSv1.2']
+        self.valid_protocols = ['TLSv1.2', 'TLSv1.1', 'TLSv1.0', 'SSLv3']
         self.session_ticket_extension = bytes([
             0x00, 0x23, 0x00, 0x00
         ])
